Adds helper module for functional tests, starts homepage test
